fix silent CLI filters, i18n output and the primary-key/integer domain
Five spec-vs-code gaps found by the previous doc review, all reproduced first and
then fixed against the code, plus two more found while verifying.

Filters that were accepted but silently ignored (exit 0, user thinks it worked):
- `ct i18n compact --dry-run` was NOT forwarded, so the "preview" really deleted
  the orphan entries (returned dict even said dry_run: False) — data loss
- `ct i18n sync --lang`, `ct i18n compact --lang`, `ct i18n status --by-table/--lang`,
  `ct gen-template --table <missing>` (silently processed 0 tables) and
  `ct export --lang <missing>` (silently exported 0 languages) had no effect
- unknown table/language values now fail with the wording already used by
  canonical_validate / run_canonical_export (UnknownTableError; case hint)

`ct i18n` output now matches the specs:
- compact printed the returned dict (`总计移除 {'dry_run': False, ...}`); the two
  documented lines were unreachable. Per-file reporting + dry-run key list added
- sync gained the summary line and a working --verbose (per-file path + counts)
- status gained the 10-cell progress bar, working --by-table, and a --json payload
  wrapped in {"langs": ...} (assembled in the CLI; the web endpoint shape is
  unchanged). Progress is translated/(total-orphan); the spec's example arithmetic
  was corrected to match that definition

Primary key type tightened to int32 (reverses the earlier "all 8 integer scalars"
decision): the export index vector is stride 4, idHash takes the low 32 bits and
the generated query signature is always ByID(int) — a uint64 primary produced an
asymmetric `ulong Id` + `ByID(int)` API and crashed the export with a TypeError
once a value exceeded int32, while `ct validate` passed it. All 19 existing
schemas use int32. Integer cells are now range-checked against their declared type
in the reader, so out-of-range values fail during parse/validate, not in the
flatbuffers builder.

Also: `ct validate` / `ct status` used to dump a Python traceback on schema-load
errors; all commands now share `_friendly_exit` ([error] + exit 1, stack goes to
the log under --verbose as the spec's escape hatch requires).

Verified: 440 passed in an isolated checkout of this commit (the working tree also
carried another session's uncommitted incremental-export work, which is
deliberately excluded — only hunks matching this change were staged).

+++ ct/tests/excel/test_canonical_template.py
@@ -141,7 +141,7 @@ def test_template_has_no_data_validation_input_prompts(tmp_path: Path) -> None:
         table="PromptFree",
         primary="Id",
         fields=[
-            FieldDef(name="Id", type="int64"),
+            FieldDef(name="Id", type="int32"),
             FieldDef(name="ItemId", type="int32", ref="Item.Id"),
             FieldDef(name="Enabled", type="bool"),
         ],
